refactor: delete Nav.tsx and rename some files
