+++ src/components/ui/Legend.tsx
