+++ src/components/ui/LevelProgressBar.tsx
